Move tests from net/instaweb/http/ to test/
