+++ pagespeed/automatic/proxy_interface_test.cc
@@ -26,9 +26,6 @@
 #include "net/instaweb/http/public/async_fetch.h"
 #include "net/instaweb/http/public/http_cache.h"
 #include "net/instaweb/http/public/logging_proto_impl.h"
-#include "net/instaweb/http/public/mock_callback.h"
-#include "net/instaweb/http/public/mock_url_fetcher.h"
-#include "net/instaweb/http/public/reflecting_test_fetcher.h"
 #include "net/instaweb/http/public/request_context.h"
 #include "net/instaweb/http/public/request_timing_info.h"
 #include "net/instaweb/rewriter/public/domain_lawyer.h"
@@ -39,6 +36,9 @@
 #include "net/instaweb/util/public/fallback_property_page.h"
 #include "net/instaweb/util/public/mock_property_page.h"
 #include "net/instaweb/util/public/property_cache.h"
+#include "test/net/instaweb/http/mock_callback.h"
+#include "test/net/instaweb/http/mock_url_fetcher.h"
+#include "test/net/instaweb/http/reflecting_test_fetcher.h"
 #include "test/net/instaweb/rewriter/rewrite_test_base.h"
 #include "test/net/instaweb/rewriter/test_rewrite_driver_factory.h"
 //#include "strings/stringpiece_utils.h"
