+++ test/net/instaweb/http/http_response_parser_test.cc
@@ -19,13 +19,13 @@
 
 #include "net/instaweb/http/public/http_response_parser.h"
 
-#include "net/instaweb/http/public/fetcher_test.h"
 #include "pagespeed/kernel/base/google_message_handler.h"
 #include "pagespeed/kernel/base/null_mutex.h"
 #include "pagespeed/kernel/base/stdio_file_system.h"
 #include "pagespeed/kernel/base/string.h"
 #include "pagespeed/kernel/base/string_writer.h"
 #include "pagespeed/kernel/http/response_headers.h"
+#include "test/net/instaweb/http/fetcher_test.h"
 #include "test/pagespeed/kernel/base/gtest.h"
 #include "test/pagespeed/kernel/base/mock_timer.h"
