+++ test/net/instaweb/http/ua_sensitive_test_fetcher_test.cc
@@ -19,10 +19,8 @@
 
 // Unit tests for UserAgentSensitiveTestFetcher
 
-#include "net/instaweb/http/public/ua_sensitive_test_fetcher.h"
+#include "test/net/instaweb/http/ua_sensitive_test_fetcher.h"
 
-#include "net/instaweb/http/public/mock_callback.h"
-#include "net/instaweb/http/public/mock_url_fetcher.h"
 #include "net/instaweb/http/public/request_context.h"
 #include "pagespeed/kernel/base/google_message_handler.h"
 #include "pagespeed/kernel/base/null_mutex.h"
@@ -36,6 +34,8 @@
 #include "pagespeed/kernel/http/request_headers.h"
 #include "pagespeed/kernel/http/response_headers.h"
 #include "pagespeed/kernel/util/platform.h"
+#include "test/net/instaweb/http/mock_callback.h"
+#include "test/net/instaweb/http/mock_url_fetcher.h"
 #include "test/pagespeed/kernel/base/gtest.h"
 #include "test/pagespeed/kernel/base/mock_timer.h"
