+++ test/net/instaweb/rewriter/rewrite_test_base.cc
@@ -30,10 +30,8 @@
 #include "net/instaweb/http/public/http_cache.h"
 #include "net/instaweb/http/public/http_cache_failure.h"
 #include "net/instaweb/http/public/http_value.h"
-#include "net/instaweb/http/public/log_record_test_helper.h"
 #include "net/instaweb/http/public/logging_proto.h"
 #include "net/instaweb/http/public/logging_proto_impl.h"
-#include "net/instaweb/http/public/mock_callback.h"
 #include "net/instaweb/http/public/request_context.h"
 #include "net/instaweb/public/global_constants.h"
 #include "net/instaweb/rewriter/cached_result.pb.h"
@@ -75,6 +73,8 @@
 #include "pagespeed/kernel/util/simple_stats.h"
 #include "pagespeed/kernel/util/url_multipart_encoder.h"
 #include "pagespeed/opt/logging/log_record.h"
+#include "test/net/instaweb/http/log_record_test_helper.h"
+#include "test/net/instaweb/http/mock_callback.h"
 #include "test/net/instaweb/rewriter/test_rewrite_driver_factory.h"
 #include "test/net/instaweb/rewriter/test_url_namer.h"
 #include "test/pagespeed/kernel/base/gtest.h"
