+++ test/net/instaweb/rewriter/test_rewrite_driver_factory.cc
@@ -25,7 +25,6 @@
 #include "base/logging.h"
 #include "net/instaweb/http/public/counting_url_async_fetcher.h"
 #include "net/instaweb/http/public/http_cache.h"
-#include "net/instaweb/http/public/mock_url_fetcher.h"
 #include "net/instaweb/http/public/rate_controller.h"
 #include "net/instaweb/http/public/rate_controlling_url_async_fetcher.h"
 #include "net/instaweb/http/public/url_async_fetcher.h"
@@ -56,6 +55,7 @@
 #include "pagespeed/kernel/util/nonce_generator.h"
 #include "pagespeed/kernel/util/platform.h"
 #include "pagespeed/kernel/util/threadsafe_lock_manager.h"
+#include "test/net/instaweb/http/mock_url_fetcher.h"
 #include "test/net/instaweb/rewriter/test_url_namer.h"
 #include "test/pagespeed/kernel/base/mem_file_system.h"
 #include "test/pagespeed/kernel/base/mock_hasher.h"
