+++ test/pagespeed/system/add_headers_fetcher_test.cc
@@ -22,8 +22,6 @@
 //
 #include "pagespeed/system/add_headers_fetcher.h"
 
-#include "net/instaweb/http/public/mock_callback.h"
-#include "net/instaweb/http/public/reflecting_test_fetcher.h"
 #include "net/instaweb/http/public/request_context.h"
 #include "net/instaweb/rewriter/public/rewrite_options.h"
 #include "pagespeed/kernel/base/google_message_handler.h"
@@ -33,6 +31,8 @@
 #include "pagespeed/kernel/http/request_headers.h"
 #include "pagespeed/kernel/http/response_headers.h"
 #include "pagespeed/kernel/util/platform.h"
+#include "test/net/instaweb/http/mock_callback.h"
+#include "test/net/instaweb/http/reflecting_test_fetcher.h"
 #include "test/net/instaweb/rewriter/rewrite_options_test_base.h"
 #include "test/pagespeed/kernel/base/gtest.h"
